added: local error truncation error, bugfix: output at wrong position
The implementation of the local truncation error and also the
convergence test was not working properly. The reason was that
at the end of the routine SolveLinearSystem(..) the timestep was updated
once more, which led to a wrong output afterwards.

The output was moved before the timestep update in order to solve
this problem.