added: Python preprocessor step
Adds a minimum working example of a python preprocessor.

In "pre_odevec.py" the pragma #ODEVEC_LU is searched and afterwards
the according decomposed values written. The basic syntax is taken
from KROME, but with sympy features used.

Within this framework "bdf_NEC.f90" is deprecated and can be removed
from the repository in a later step. It is generated from
"bdf_NEC.F90" which serves as a template.

Until now still only for dense systems.